+++ src/main/java/com/project/sfoc/redis/RefreshTokenRepository.java
@@ -1,4 +1,4 @@
-package com.project.sfoc.security.jwt.redis;
+package com.project.sfoc.redis;
 
 import org.springframework.data.repository.CrudRepository;
