+++ src/main/java/com/project/sfoc/redis/RefreshTokenService.java
@@ -1,4 +1,4 @@
-package com.project.sfoc.security.jwt.redis;
+package com.project.sfoc.redis;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
