+++ src/main/java/com/project/sfoc/security/handler/CustomLogoutHandler.java
@@ -1,7 +1,7 @@
-package com.project.sfoc.security;
+package com.project.sfoc.security.handler;
 
-import com.project.sfoc.security.jwt.JwtUtil;
-import com.project.sfoc.security.jwt.redis.RefreshTokenService;
+import com.project.sfoc.security.HttpUtil;
+import com.project.sfoc.redis.RefreshTokenService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
@@ -10,27 +10,24 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.logout.LogoutHandler;
 import org.springframework.stereotype.Component;
 
-import java.util.Optional;
-
 @Component
 @RequiredArgsConstructor
 @Slf4j
 public class CustomLogoutHandler implements LogoutHandler {
 
-    private final JwtUtil jwtUtil;
+    private final HttpUtil httpUtil;
     private final RefreshTokenService refreshTokenService;
 
     @Override
     public void logout(HttpServletRequest request, HttpServletResponse response, Authentication authentication) {
-        Optional<String> refreshToken = jwtUtil.resolveRefreshToken(request);
+        String refreshToken = httpUtil.resolveRefreshToken(request);
 
-        refreshToken.filter((token) -> jwtUtil.isNotExpired(token) && refreshTokenService.isValid(token))
-                .ifPresentOrElse(refreshTokenService::deleteRefreshToken, this::handleLogoutFailure);
+        if (!refreshTokenService.isValid(refreshToken)) {
+            throw new RuntimeException("일치하는 refresh token이 없음");
+        }
 
+        refreshTokenService.deleteRefreshToken(refreshToken);
         log.info("logout 완료");
     }
 
-    private void handleLogoutFailure() {
-        throw new RuntimeException("일치하는 refresh token이 없습니다.");
-    }
 }
