+++ src/main/java/com/project/sfoc/security/jwt/JwtUtil.java
@@ -1,25 +1,18 @@
 package com.project.sfoc.security.jwt;
 
-import com.project.sfoc.security.jwt.redis.RefreshToken;
-import com.project.sfoc.security.jwt.redis.RefreshTokenRepository;
-import io.jsonwebtoken.ExpiredJwtException;
+import com.project.sfoc.redis.RefreshToken;
+import com.project.sfoc.redis.RefreshTokenRepository;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.jackson.io.JacksonDeserializer;
 import io.jsonwebtoken.lang.Maps;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.ResponseCookie;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.util.UriComponentsBuilder;
 
 import javax.crypto.SecretKey;
 import javax.crypto.spec.SecretKeySpec;
 import java.nio.charset.StandardCharsets;
-import java.util.Arrays;
 import java.util.Date;
-import java.util.Optional;
 
 @Component
 public class JwtUtil {
@@ -29,10 +22,7 @@ public class JwtUtil {
     private static final String ISSUER = "sfoc";
     private static final String TOKEN_TYPE = "token_type";
     private static final String USER_INFO = "user_info";
-    private static final String TOKEN_PARAM = "token";
-    private static final String CALLBACK_URL = "https://sfoc-scheduler.github.io/SFOC-frontend";
 
-    public static final String ACCESS_HEADER = "Authorization";
     public static final String ACCESS_TYPE = "AccessToken";
     public static final String REFRESH_TYPE = "RefreshToken";
 
@@ -82,20 +72,6 @@ public class JwtUtil {
         return refreshToken;
     }
 
-    public boolean isNotExpired(String token) {
-        try {
-            return !Jwts.parser()
-                    .verifyWith(secretKey)
-                    .build()
-                    .parseSignedClaims(token)
-                    .getPayload()
-                    .getExpiration()
-                    .before(new Date());
-        } catch (ExpiredJwtException e) {
-            return false;
-        }
-    }
-
     public UserInfo getUserInfo(String token) {
         return Jwts.parser()
                 .verifyWith(secretKey)
@@ -104,38 +80,6 @@ public class JwtUtil {
                 .parseSignedClaims(token)
                 .getPayload()
                 .get(USER_INFO, UserInfo.class);
-    }
-
-    public String createCookie(String name, String value) {
-        return ResponseCookie.from(name, value)
-                .httpOnly(true)
-                .secure(true)
-                .sameSite("None")
-                .path("/")
-                .maxAge(1_209_600)  // 2주
-                .build()
-                .toString();
-    }
-
-    public Optional<String> resolveAccessToken(HttpServletRequest request) {
-        return Optional.ofNullable(request.getHeader(ACCESS_HEADER))
-                .filter(authorization -> authorization.startsWith("Bearer "))
-                .map(authorization -> authorization.split(" ")[1]);
-    }
-
-    public Optional<String> resolveRefreshToken(HttpServletRequest request) {
-        return Optional.ofNullable(request.getCookies())
-                .flatMap(cookies -> Arrays.stream(cookies)
-                        .filter(cookie -> cookie.getName().equals(REFRESH_TYPE))
-                        .findFirst()
-                        .map(Cookie::getValue));
-    }
-
-    public String getRedirectUrl(String accessToken) {
-        return UriComponentsBuilder.fromUriString(CALLBACK_URL)
-                .queryParam(TOKEN_PARAM, accessToken)
-                .build()
-                .toUriString();
     }
 
 }
