+++ src/main/java/com/project/sfoc/security/oauth2/OAuth2UserInfo.java
@@ -1,4 +1,4 @@
-package com.project.sfoc.security;
+package com.project.sfoc.security.oauth2;
 
 import com.project.sfoc.entity.Provider;
 import com.project.sfoc.entity.user.User;
